refactor(explore): split 889-line monolith into tab-pluggable modules
Finding #6 from the audit, escalated because the curated catalog
(use-case packs, editorial collections, bridges to YouTube / Reddit)
is the next investment area. A flat 889-line file is the wrong shape
to add 3+ new tab kinds to.

Extracted:

  src/components/explore/
    feed-row.tsx              — shared row (favicon, preview, add/remove)
    category-grid.tsx         — tile grid + GridItem type
    category-detail.tsx       — feed list + bulk add/remove all
    featured-tab.tsx          — curated landing
    topics-tab.tsx            — section → subcategory → feeds
    countries-tab.tsx         — country → feeds
    search-results-view.tsx   — grouped hits across all tabs
    explore-catalog.tsx       — shell only (header, search, tabs, listbox)

Pure refactor — no behaviour change, no public API change.
explore-catalog.tsx is now 317 lines (header + search input + tab
switch + keyboard nav + listbox container). The TAB_DESCRIPTORS list
makes adding a new tab a three-touch operation: create the file,
extend the BrowseTab union, add a descriptor + render branch.

17 catalog tests stay green, full suite 2475 passing.
